+++ src/formapplications/components/NumberInput.svelte
@@ -1,6 +1,6 @@
 <script>
-	import { localize } from '#runtime/util/i18n';
-  import { applyStyles } from '#runtime/svelte/action/dom/style';
+	let localize = game.i18n.localize.bind(game.i18n);
+	import { applyStyles } from "./actions.svelte.js"
   import { writable } from "svelte/store";
 
   export let setting;