+++ src/formapplications/effects-ui/Tabs.svelte
@@ -1,5 +1,5 @@
 <script>
-  import { localize } from '#runtime/util/i18n';
+  let localize = game.i18n.localize.bind(game.i18n);
 
   export let activeTab;
   export let tabs;